+++ tools/extraChart/tool.js
@@ -196,7 +196,7 @@
 
     pieceTension = (x, y, board, withSupport) => {
       const underAttack = [];
-      var ch = board[y][x];
+      const ch = board[y][x];
       if (!ch) return underAttack;
       const m = ch === ch.toUpperCase() ? 1 : -1;
       switch (ch.toLowerCase()) {
@@ -478,7 +478,9 @@
           underAttack.push.apply(underAttack, ua);
         }
       }
-      result = [...new Set(underAttack.map(i => i.x + ',' + i.y + '=' + i.pc))].map(i => i.split('=')[1].toLowerCase()).reduce((acc, val) => this.pieceMaterial[val] + acc, 0);
+      result = [...new Set(underAttack.map(i => i.x + ',' + i.y + '=' + i.pc))]
+                  .map(i => i.split('=')[1].toLowerCase())
+                  .reduce((acc, val) => this.pieceMaterial[val] + acc, 0);
       this.tensionCache.set(fen,result);
       return result;
     };
@@ -524,9 +526,13 @@
       return !!moves.length;
     };
 
+    materialWonCache = new LiChessTools.MaxSizedMap(10000);
     materialWon = (board, x, y) => {
       const lt = this.lichessTools;
       const Math = lt.global.Math;
+      const key = [lt.getFenFromBoard(board),x,y].join('|');
+      let result = this.materialWonCache.get(key);
+      if (result) return result;
       board = lt.clone(board);
       const ch = board[y][x];
       if (!ch) return 0;
@@ -536,16 +542,22 @@
         .sort((m1, m2) => this.pieceMaterial[m1.spc.toLowerCase()] - this.pieceMaterial[m2.spc.toLowerCase()]);
       const move = moves[0];
       if (!move) return 0;
-      let result = this.pieceMaterial[ch.toLowerCase()] * side; //capture piece
+      result = this.pieceMaterial[ch.toLowerCase()] * side; //capture piece
       board[y][x] = move.spc;
       board[move.sy][move.sx] = null;
       result += this.materialWon(board, x, y);
-      return Math.sign(result) != side ? 0 : result;
+      result = Math.sign(result) != side ? 0 : result;
+      this.materialWonCache.set(key,result);
+      return result;
     };
 
+    maxMaterialWonCache = new LiChessTools.MaxSizedMap(10000);
     maxMaterialWon = (board, m) => {
       const lt = this.lichessTools;
       const Math = lt.global.Math;
+      const key = [lt.getFenFromBoard(board),m].join('|');
+      let result = this.maxMaterialWonCache.get(key);
+      if (result) return result;
       let mx = 0;
       for (let y = 0; y < 8; y++) {
         for (let x = 0; x < 8; x++) {
@@ -559,7 +571,9 @@
           }
         }
       }
-      return mx * m;
+      result = mx * m;
+      this.maxMaterialWonCache.set(key, result);
+      return result;
     };
 
     smooth = (points) => {
@@ -828,12 +842,17 @@
       return (w1 - w2) * side;
     }
 
+    computeBrilliantCache = new LiChessTools.MaxSizedMap(1000);
     computeBrilliant = (side, node, prevNode, prev2Node) => {
       const lt = this.lichessTools;
       const Math = lt.global.Math;
       const cp1 = this.getCp(this.getNodeCeval(node));
       const cp2 = this.getCp(this.getNodeCeval(prevNode));
       if ([cp1,cp2].findIndex(cp=>!cp && cp!==0)>=0) return 0;
+
+      const key = [node?.fen, side, cp1, cp2, prevNode?.fen, prev2Node?.fen].join('|');
+      let result = this.computeBrilliantCache.get(key);
+      if (result) return result;
 
       const threshold = Math.abs(cp1) > 200 || this.hasTacticalMotif(node, side, prevNode) ? -50 : -25;
       if ((cp1 - cp2) * side < threshold) return 0;
@@ -876,7 +895,9 @@
       board = lt.getBoardFromFen(prev2Node.fen);
       const mmw3 = this.maxMaterialWon(board, side) / 100;
       const bril = (mmw3 - mmw1 - delta);
-      return bril + bonus;
+      result = bril + bonus;
+      this.computeBrilliantCache.set(key,result);
+      return result;
     };
 
     computeSpectacle = (board, x, y, side) => {
@@ -1323,42 +1344,42 @@
       const fill = (container, count, color) => {
         let elem = $('div.lichessTools-goodMoves', container);
         if (!elem.length) {
-          const adviceSummary = $('div.advice-summary__player', container);
-          if (adviceSummary.length) {
-          elem = $('<div></div>')
-            .addClass('lichessTools-goodMoves')
-            .addClass('advice-summary__error')
-            .text(' ' + trans.noarg('goodMovesText'))
-            .prepend($('<strong></strong>'))
-            .attr('title', trans.noarg('goodMovesTitle'))
-            .on('click', (ev) => {
-              ev.preventDefault();
-              lt.jumpToGlyphSymbols(this.options.moreBrilliant ? ['!?', '!!'] : ['!', '!?', '!!', lt.icon.WhiteStar], color);
-            })
-            .on('pointermove', (ev) => {
-              const chart = this._chart;
-              if (!chart) return;
-              const dataset = chart.data.datasets.at(0);
-              if (!dataset) return;
-              const elems = this.getInterestingMoveElements(color);
-              if (!this.colors.beforeInterestingMoves) this.colors.beforeInterestingMoves = dataset.hoverBackgroundColor || this.colors.originalChart;
-              dataset.hoverBackgroundColor = this.colors.interestingMoves();
-              dataset.pointHoverBackgroundColor = this.colors.interestingMoves();
-              this.safeSetActiveElements(chart,elems,dataset);
-              chart.update('none');
-            })
-            .on('pointerleave', (ev) => {
-              const chart = this._chart;
-              if (!chart) return;
-              const dataset = chart.data.datasets.at(0);
-              if (!dataset) return;
-              const elems = [];
-              dataset.hoverBackgroundColor = this.colors.beforeInterestingMoves;
-              dataset.pointHoverBackgroundColor = this.colors.beforeInterestingMoves;
-              this.safeSetActiveElements(chart,elems,dataset);
-              chart.update('none');
-            })
-            .insertAfter(adviceSummary);
+          const adviceAcpl = $('div.advice-summary__acpl', container);
+          if (adviceAcpl.length) {
+            elem = $('<div></div>')
+              .addClass('lichessTools-goodMoves')
+              .addClass('advice-summary__error')
+              .text(' ' + trans.noarg('goodMovesText'))
+              .prepend($('<strong></strong>'))
+              .attr('title', trans.noarg('goodMovesTitle'))
+              .on('click', (ev) => {
+                ev.preventDefault();
+                lt.jumpToGlyphSymbols(this.options.moreBrilliant ? ['!?', '!!'] : ['!', '!?', '!!', lt.icon.WhiteStar], color);
+              })
+              .on('pointermove', (ev) => {
+                const chart = this._chart;
+                if (!chart) return;
+                const dataset = chart.data.datasets.at(0);
+                if (!dataset) return;
+                const elems = this.getInterestingMoveElements(color);
+                if (!this.colors.beforeInterestingMoves) this.colors.beforeInterestingMoves = dataset.hoverBackgroundColor || this.colors.originalChart;
+                dataset.hoverBackgroundColor = this.colors.interestingMoves();
+                dataset.pointHoverBackgroundColor = this.colors.interestingMoves();
+                this.safeSetActiveElements(chart,elems,dataset);
+                chart.update('none');
+              })
+              .on('pointerleave', (ev) => {
+                const chart = this._chart;
+                if (!chart) return;
+                const dataset = chart.data.datasets.at(0);
+                if (!dataset) return;
+                const elems = [];
+                dataset.hoverBackgroundColor = this.colors.beforeInterestingMoves;
+                dataset.pointHoverBackgroundColor = this.colors.beforeInterestingMoves;
+                this.safeSetActiveElements(chart,elems,dataset);
+                chart.update('none');
+              })
+              .prependTo(adviceAcpl);
           }
         }
         if (elem.length) {
@@ -1477,10 +1498,11 @@
       const lichess = lt.lichess;
       const $ = lt.$;
       const trans = lt.translator;
+      const analysis = lichess.analysis;
 
       if (!forced && lt.random() > 0.9) forced = true; // hack to sometimes update this anyway
 
-      if (!lichess.analysis) return;
+      if (!analysis) return;
 
       const currentBrilliant = [this.options.brilliant, this.options.moreBrilliant].join(',');
       if (!forced && this.prevBrilliant != currentBrilliant) {
@@ -1490,13 +1512,15 @@
       let localLine = null;
       if (this.options.local) {
         localLine = this.getLocalLine();
+      } else {
+        localLine = analysis.mainline;
       }
 
-      if (lichess.analysis.mainline.find(n => n.eval && n.eval._originator != 'lichessTools')) {
+      if (analysis.mainline.find(n => n.eval && n.eval._originator != 'lichessTools')) {
         $('#acpl-chart-container').removeClass('lichessTools-extraChart');
         $('form.future-game-analysis').remove();
       }
-      if (!lichess.analysis.mainline.find(n => this.getNodeCeval(n))) {
+      if (!analysis.mainline.find(n => this.getNodeCeval(n))) {
         if (!localLine || !localLine.find(n => this.getNodeCeval(n))) {
           $('#acpl-chart-container.lichessTools-extraChart').remove();
           this.clearCharts();
@@ -1508,7 +1532,7 @@
 
       let container = $.cached('#acpl-chart-container.lichessTools-extraChart, div.computer-analysis.active #acpl-chart-container, div.study__server-eval.ready',3000);
       if (!chart) {
-        const serverEval = lichess.analysis.study?.serverEval;
+        const serverEval = analysis.study?.serverEval;
         chart = /*serverEval?.requested && */serverEval?.chart;
         if (!chart) {
           const canvas = $('canvas#acpl-chart', container)[0];
@@ -1533,7 +1557,7 @@
               container.insertBefore(message);
               loadChartModule = true;
             } else
-            if (lichess.analysis.study) {
+            if (analysis.study) {
               const elem = $('.study__message');
               if (elem.length && !elem.closest('.study__glyphs').length) {
                 elem.append(container)
@@ -1547,8 +1571,8 @@
             }
             if (loadChartModule) {
               const mod = await this.getChartModule();
-              const mainline = lichess.analysis.mainline;
-              chart = await mod.acpl(container.find("#acpl-chart")[0], lichess.analysis.data, mainline, lichess.analysis.trans);
+              const mainline = analysis.mainline;
+              chart = await mod.acpl(container.find("#acpl-chart")[0], analysis.data, mainline, analysis.trans);
             }
           }
         }
@@ -1729,7 +1753,7 @@
           chart.data.datasets.push({
             label: 'Material',
             type: 'line',
-            data: this.smooth(this.getMaterialData(lichess.analysis.mainline)),
+            data: this.smooth(this.getMaterialData(analysis.mainline)),
             borderWidth: 2,
             borderDash: [3, 3],
             cubicInterpolationMode: this.options.smooth ? 'monotone' : 'default',
@@ -1745,8 +1769,8 @@
           updateChart = true;
         } else {
           const dataset = chart.data.datasets[existingMaterial];
-          if (dataset.data?.length < lichess.analysis.mainline.length) {
-            const newData = this.smooth(this.getMaterialData(lichess.analysis.mainline));
+          if (dataset.data?.length < analysis.mainline.length) {
+            const newData = this.smooth(this.getMaterialData(analysis.mainline));
             dataset.data = newData;
             updateChart = true;
           }
@@ -1764,7 +1788,7 @@
           chart.data.datasets.push({
             label: 'Principled',
             type: 'line',
-            data: this.smooth(this.getPrincipledData(lichess.analysis.mainline)),
+            data: this.smooth(this.getPrincipledData(analysis.mainline)),
             borderWidth: 2,
             borderDash: [4, 2],
             cubicInterpolationMode: this.options.smooth ? 'monotone' : 'default',
@@ -1780,8 +1804,8 @@
           updateChart = true;
         } else {
           const dataset = chart.data.datasets[existingPrincipled];
-          if (dataset.data?.length < lichess.analysis.mainline.length) {
-            const newData = this.smooth(this.getPrincipledData(lichess.analysis.mainline));
+          if (dataset.data?.length < analysis.mainline.length) {
+            const newData = this.smooth(this.getPrincipledData(analysis.mainline));
             dataset.data = newData;
             updateChart = true;
           }
@@ -1869,7 +1893,7 @@
         updateChart = true;
       }
       if (this.options.accuracy || this.tempAccuracyOn) {
-        const mainline = lichess.analysis.mainline;
+        const mainline = analysis.mainline;
         if (existingAccuracy < 0) {
           chart.data.datasets.push({
             label: 'Accuracy',
@@ -1904,7 +1928,7 @@
         updateChart = true;
       }
       if (this.options.sharpness) {
-        const mainline = lichess.analysis.mainline;
+        const mainline = analysis.mainline;
         if (existingSharpness < 0) {
           chart.data.datasets.push({
             label: 'Sharpness',
@@ -1939,7 +1963,7 @@
         updateChart = true;
       }
       if (this.options.coord) {
-        const mainline = lichess.analysis.mainline;
+        const mainline = analysis.mainline;
         if (existingCoord < 0) {
           chart.data.datasets.push({
             label: 'Coordination',
@@ -2009,7 +2033,7 @@
       }
       if (this.options.tension) {
         if (existingMaxTension < 0) {
-          const x = this.getMaxTension(lichess.analysis.mainline);
+          const x = this.getMaxTension(analysis.mainline);
           chart.data.datasets.push({
             label: 'Max tension',
             type: 'line',
@@ -2035,8 +2059,8 @@
         } else {
           const dataset = chart.data.datasets[existingMaxTension];
           const existingData = dataset.data;
-          if (this.prevMainlineLength != lichess.analysis.mainline.length) {
-            const x = this.getMaxTension(lichess.analysis.mainline);
+          if (this.prevMainlineLength != analysis.mainline.length) {
+            const x = this.getMaxTension(analysis.mainline);
             dataset.data = [
               { x: x, y: -1.05 },
               { x: x, y: 1.05 }
@@ -2054,7 +2078,7 @@
       }
       if (this.options.potential) {
         if (existingMaxPotential < 0) {
-          const x = this.getMaxPotential(lichess.analysis.mainline);
+          const x = this.getMaxPotential(analysis.mainline);
           chart.data.datasets.push({
             label: 'Max potential',
             type: 'line',
@@ -2079,8 +2103,8 @@
           updateChart = true;
         } else {
           const dataset = chart.data.datasets[existingMaxPotential];
-          if (this.prevMainlineLength != lichess.analysis.mainline.length) {
-            const x = this.getMaxPotential(lichess.analysis.mainline);
+          if (this.prevMainlineLength != analysis.mainline.length) {
+            const x = this.getMaxPotential(analysis.mainline);
             dataset.data = [
               { x: x, y: -1.05 },
               { x: x, y: 1.05 }
@@ -2096,14 +2120,15 @@
         middleGameText = 'Middlegame';
       }
       let existingMiddleGame = chart.data.datasets.findIndex(s => s.label === middleGameText);
-      if (existingMiddleGame >= 0 && !this.options.local) {
+      // do not delete the original middle game line
+      /*if (existingMiddleGame >= 0 && !this.options.local) {
         chart.data.datasets.splice(existingMiddleGame, 1);
         existingMiddleGame = -1;
         updateChart = true;
-      }
+      }*/
       if (this.options.local) {
         if (existingMiddleGame < 0) {
-          const x = this.getGamePhases(lichess.analysis.mainline).middleGame;
+          const x = this.getGamePhases(analysis.mainline).middleGame;
           if (x) {
           chart.data.datasets.push({
             label: middleGameText,
@@ -2123,7 +2148,7 @@
               offset: -5,
               align: 45,
               rotation: 90,
-              formatter: _ => trans.noarg('middleGameTitle')
+              formatter: _ => middleGameText
             }
           });
           } else {
@@ -2135,8 +2160,8 @@
           updateChart = true;
         } else {
           const dataset = chart.data.datasets[existingMiddleGame];
-          if (this.prevMainlineLength != lichess.analysis.mainline.length) {
-            const x = this.getGamePhases(lichess.analysis.mainline).middleGame;
+          if (this.prevMainlineLength != analysis.mainline.length) {
+            const x = this.getGamePhases(analysis.mainline).middleGame;
             dataset.data = [
               { x: x, y: -1.05 },
               { x: x, y: 1.05 }
@@ -2153,14 +2178,15 @@
       }
 
       let existingEndGame = chart.data.datasets.findIndex(s => s.label === endGameText);
-      if (existingEndGame >= 0 && !this.options.local) {
+      // do not delete the original endgame line
+      /*if (existingEndGame >= 0 && !this.options.local) {
         chart.data.datasets.splice(existingEndGame, 1);
         existingEndGame = -1;
         updateChart = true;
-      }
+      }*/
       if (this.options.local) {
         if (existingEndGame < 0) {
-          const x = this.getGamePhases(lichess.analysis.mainline).endGame;
+          const x = this.getGamePhases(analysis.mainline).endGame;
           if (x) {
           chart.data.datasets.push({
             label: endGameText,
@@ -2180,7 +2206,7 @@
               offset: -5,
               align: 45,
               rotation: 90,
-              formatter: _ => trans.noarg('endGameTitle')
+              formatter: _ => endGameText
             }
           });
           } else {
@@ -2192,8 +2218,8 @@
           updateChart = true;
         } else {
           const dataset = chart.data.datasets[existingEndGame];
-          if (this.prevMainlineLength != lichess.analysis.mainline.length) {
-            const x = this.getGamePhases(lichess.analysis.mainline).endGame;
+          if (this.prevMainlineLength != analysis.mainline.length) {
+            const x = this.getGamePhases(analysis.mainline).endGame;
             dataset.data = [
               { x: x, y: -1.05 },
               { x: x, y: 1.05 }
@@ -2204,7 +2230,7 @@
       }
 
       if (this.options.brilliant) {
-        this.setBrilliant(lichess.analysis.mainline, forced);
+        this.setBrilliant(analysis.mainline, forced);
 
         this.showGoodMoves(forced);
       }
@@ -2216,7 +2242,7 @@
       const maxX = Math.max.apply(null, chart.data.datasets
         .filter(d => d.type == 'line')
         .map(d => d.data?.at(-1)?.x)
-        .concat([lichess.analysis.mainline.at(-1)?.ply]));
+        .concat([analysis.mainline.at(-1)?.ply]));
       if (maxX != chart.options?.scales?.x?.max) {
         chart.options.scales.x.max = maxX;
         updateChart = true;
@@ -2224,7 +2250,7 @@
 
       if (updateChart) {
         chart.update('none');
-        this.prevMainlineLength = lichess.analysis.mainline.length;
+        this.prevMainlineLength = analysis.mainline.length;
       }
       this.prevSmooth = this.options.smooth;
     };
@@ -2409,12 +2435,13 @@
 
     showAccuracy = (ev)=>{
       const el = ev.currentTarget;
+      if (!el?.offsetParent) return;
       const lt = this.lichessTools;
       const lichess = lt.lichess;
       const $ = lt.$;
       const trans = lt.translator;
-      var isWhite = $(el).closest('.advice-summary__side').has('.is.white').length ? 1 : 0;
-      var localLine = this.getLocalLine().filter(n=>n.ply && (n.ply % 2) == isWhite);
+      const isWhite = $(el).closest('.advice-summary__side').has('.is.white').length ? 1 : 0;
+      const localLine = this.getLocalLine().filter(n=>n.ply && (n.ply % 2) == isWhite);
       if (!localLine.length) return;
       const dict = new Map();
       for (const node of localLine) {